Inventory Implementation

+++ Pavas.Application/Executors/Inventory/Commands/Add/AppAddInventoryCommand.cs
@@ -10,5 +10,5 @@ public record AppAddInventoryCommand(
     int CompanyId,
     InventoryType Type,
     decimal Price,
-    int Quantity
+    decimal Quantity
 ) : ICommand;
